pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "assignment2"
version = "1.0.0"
description = "CIS 6930 Assignment 2 - RAG System"
authors = [{name = "Your Name", email = "your.email@ufl.edu"}]
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3",
    "langchain-community>=0.3",
    "langchain-openai>=0.2",
    "chromadb>=0.5",
    "sentence-transformers>=3.0",
    "pypdf>=4.0",
    "python-dotenv>=1.0",
    "loguru>=0.7",
    "pytest>=8.0",
]

[project.scripts]
rag-index = "index:main"
rag-query = "query:main"
rag-evaluate = "run_evaluation:main"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
